feat(kmp): migrate objects/cells/ and sim/ tests to commonTest (5 files)
Migrate 5 test files from jvmTest to commonTest for cross-platform
coverage on both JVM and linuxX64:

- CellConnectionTest: extracted 3 @Nested classes to top-level,
  replaced Math.abs/Math.max with kotlin.math equivalents
- DynamicRailSemaphoreTest: extracted 2 @Nested classes to top-level
  (EqualsContractTest, ListenerThreadSafetyTest)
- RailSwitchTest: flattened 4 @Nested classes into single class,
  converted assertThrows to assertFailsWith
- TimeTest: extracted 6 @Nested classes to 6 top-level test classes
- SimpleIntegrationTest: extracted 4 @Nested classes to top-level,
  kDisco Variable import is KMP-safe

All conversions: JUnit5 @Test -> kotlin.test @Test,
@BeforeEach -> @BeforeTest, removed @DisplayName/@Nested

JVM: 1624 tests passing, Native: 245 tests passing
