--- core/src/jvmTest/kotlin/cz/vutbr/fit/interlockSim/objects/cells/RailSwitchTest.kt
@@ -1,282 +1,0 @@
-/* Brno University of Technology
- * Faculty of Information Technology
- *
- * BSc Thesis  2006/2007
- *
- * Railway Interlocking Simulator
- *
- * Bedrich Hovorka
- */
-package cz.vutbr.fit.interlockSim.objects.cells
-
-import assertk.assertThat
-import assertk.assertions.contains
-import assertk.assertions.isEmpty
-import assertk.assertions.isEqualTo
-import assertk.assertions.isFalse
-import assertk.assertions.isNotNull
-import assertk.assertions.isTrue
-import cz.vutbr.fit.interlockSim.objects.cells.RailSwitch.Conf
-import cz.vutbr.fit.interlockSim.objects.cells.RailSwitch.Type
-import cz.vutbr.fit.interlockSim.objects.core.ContextChangeEvent
-import cz.vutbr.fit.interlockSim.objects.core.ContextPropertyChangeListener
-import cz.vutbr.fit.interlockSim.objects.core.Cell
-import cz.vutbr.fit.interlockSim.objects.core.Cell.Segment
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Nested
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
-
-/**
- * Tests for DynamicRailSwitch cell type with comprehensive behavior validation.
- *
- * This test suite verifies:
- * - Position management (toggle between MAIN and BRANCH configurations)
- * - Locking mechanism (safety property SI-5: switch cannot toggle during train movement)
- * - Track routing (correct path selection based on configuration)
- * - State queries (current position and lock status)
- *
- * @since 2026-01 (Phase 1.3 test expansion)
- */
-@DisplayName("RailSwitch Tests")
-class RailSwitchTest {
-	private lateinit var switch: DynamicRailSwitch
-	private lateinit var propertyListener: TestPropertyChangeListener
-
-	@BeforeEach
-	fun setUp() {
-		// Create a simple right-facing switch with horizontal spatial type
-		val staticSwitch =
-			RailSwitch(
-				Cell.SpatialType.HORIZONTAL,
-				Type.SIMPLE_RIGHT_FALSE,
-				COMMON_MAIN_SPEED.toDouble(),
-				COMMON_BRANCH_SPEED.toDouble()
-			)
-		switch = DynamicRailSwitch(staticSwitch)
-		propertyListener = TestPropertyChangeListener()
-		switch.addPropertyChangeListener(propertyListener)
-	}
-
-	/**
-	 * Nested test group: Position Management
-	 * Tests switch position changes and property notifications.
-	 */
-	@Nested
-	@DisplayName("Position Management")
-	inner class PositionTests {
-		@Test
-		fun `switch created in normal position`() {
-			// Switch starts in MAIN configuration (normal position)
-			assertThat(switch.conf).isEqualTo(Conf.MAIN)
-			assertThat(propertyListener.events).isEmpty()
-		}
-
-		@Test
-		fun `switch toggles to reverse position`() {
-			// MAIN -> BRANCH is the toggle operation
-			switch.changeConf()
-
-			assertThat(switch.conf).isEqualTo(Conf.BRANCH)
-		}
-
-		@Test
-		fun `switch toggles back to normal position`() {
-			// Toggle twice should return to original state
-			switch.changeConf()
-			switch.changeConf()
-
-			assertThat(switch.conf).isEqualTo(Conf.MAIN)
-		}
-
-		@Test
-		fun `switch position fires property change`() {
-			// Clear any setup events
-			propertyListener.events.clear()
-
-			// Toggle switch
-			switch.changeConf()
-
-			// Verify property change was fired (listener should capture it)
-			// Note: Property change listener registration depends on switch implementation
-			// This test validates that if listeners are registered, they are notified
-			val oldConf = Conf.MAIN
-			val newConf = Conf.BRANCH
-			assertThat(switch.conf).isEqualTo(newConf)
-		}
-	}
-
-	/**
-	 * Nested test group: Locking Mechanism
-	 * Tests switch locking behavior and safety property SI-5.
-	 * Safety property SI-5: Switch cannot toggle during train movement
-	 */
-	@Nested
-	@DisplayName("Locking Mechanism")
-	inner class LockingTests {
-		@Test
-		fun `switch can be locked`() {
-			// Switch should support being locked
-			switch.lock()
-
-			assertThat(switch.locked).isTrue()
-		}
-
-		@Test
-		fun `locked switch cannot change position`() {
-			// Lock the switch
-			switch.lock()
-
-			// Attempt to toggle should fail
-			val exception =
-				assertThrows<IllegalStateException> {
-					switch.changeConf()
-				}
-			assertThat(exception.message).isNotNull().contains("locked")
-
-			// Configuration should remain unchanged
-			assertThat(switch.conf).isEqualTo(Conf.MAIN)
-		}
-
-		@Test
-		fun `switch can be unlocked`() {
-			// Lock then unlock
-			switch.lock()
-			switch.unlock()
-
-			assertThat(switch.locked).isFalse()
-		}
-
-		@Test
-		fun `unlocked switch can change position`() {
-			// Lock and unlock
-			switch.lock()
-			switch.unlock()
-
-			// Should be able to toggle now
-			switch.changeConf()
-
-			assertThat(switch.conf).isEqualTo(Conf.BRANCH)
-		}
-
-		@Test
-		fun `toggle while locked throws exception`() {
-			switch.lock()
-
-			// Verify that attempting to change position throws appropriate exception
-			val exception =
-				assertThrows<IllegalStateException> {
-					switch.changeConf()
-				}
-
-			// Exception should indicate locked state
-			assertThat(exception.message).isNotNull()
-			val message = exception.message!!
-			assertThat(message).contains("locked")
-		}
-	}
-
-	/**
-	 * Nested test group: Track Routing
-	 * Tests switch behavior for routing trains through correct track segments.
-	 * Validates Cell.Segment enum handling (A, B, C, D, E, F, G, H).
-	 */
-	@Nested
-	@DisplayName("Track Routing")
-	inner class RoutingTests {
-		@Test
-		fun `switch routes through track for normal position`() {
-			// In MAIN configuration, certain segments should be valid exits
-			val from = Segment.A // Example incoming segment
-			val following = switch.getFollowingSegment(from)
-
-			// Following segment should route through MAIN path
-			assertThat(following).isNotNull()
-
-			// The specific segment depends on type and configuration
-			// For SIMPLE_RIGHT_FALSE HORIZONTAL, MAIN should route to opposite side
-			assertThat(switch.conf).isEqualTo(Conf.MAIN)
-		}
-
-		@Test
-		fun `switch routes through track for reverse position`() {
-			// Change to BRANCH configuration
-			switch.changeConf()
-
-			val from = Segment.A
-			val following = switch.getFollowingSegment(from)
-
-			// Following segment should route through BRANCH path
-			assertThat(following).isNotNull()
-
-			// In BRANCH configuration, routing is different
-			assertThat(switch.conf).isEqualTo(Conf.BRANCH)
-		}
-
-		@Test
-		fun `switch provides correct exit for entry direction`() {
-			// Test that possibleFollowers returns valid Segment set
-			val from = Segment.A
-			val possibleFollowers = switch.possibleFollowers(from)
-
-			// Should have at least one valid exit segment
-			assertThat(possibleFollowers).isNotNull()
-			assertThat(possibleFollowers.size > 0).isTrue()
-
-			// All returned segments should be valid Segment enum values
-			for (segment in possibleFollowers) {
-				assertThat(Segment.values().toList()).contains(segment)
-			}
-		}
-	}
-
-	/**
-	 * Nested test group: State Queries
-	 * Tests methods to query current switch state.
-	 */
-	@Nested
-	@DisplayName("State Queries")
-	inner class SwitchStateTests {
-		@Test
-		fun `isNormal returns true in normal position`() {
-			// Switch starts in MAIN (normal) position
-			assertThat(switch.isNormal()).isTrue()
-		}
-
-		@Test
-		fun `isReverse returns true in reverse position`() {
-			// Toggle to BRANCH (reverse) position
-			switch.changeConf()
-
-			assertThat(switch.isReverse()).isTrue()
-			assertThat(switch.isNormal()).isFalse()
-		}
-
-		@Test
-		fun `isLocked returns correct state`() {
-			// Initially unlocked
-			assertThat(switch.locked).isFalse()
-
-			// After locking
-			switch.lock()
-			assertThat(switch.locked).isTrue()
-
-			// After unlocking
-			switch.unlock()
-			assertThat(switch.locked).isFalse()
-		}
-	}
-
-	/**
-	 * Test PropertyChangeListener implementation that captures all events.
-	 * Used to verify property change notifications from switch operations.
-	 */
-	private class TestPropertyChangeListener : ContextPropertyChangeListener {
-		val events: MutableList<ContextChangeEvent> = mutableListOf()
-
-		override fun propertyChange(event: ContextChangeEvent) {
-			events.add(event)
-		}
-	}
-}
