--- core/src/jvmTest/kotlin/cz/vutbr/fit/interlockSim/sim/SimpleIntegrationTest.kt
@@ -1,405 +1,0 @@
-/* Brno University of Technology
- * Faculty of Information Technology
- *
- * BSc Thesis  2006/2007
- *
- * Railway Interlocking Simulator
- *
- * Unit tests for SimpleIntegration
- * Phase 2.4 test implementation - 2025
- */
-package cz.vutbr.fit.interlockSim.sim
-
-import assertk.assertThat
-import assertk.assertions.isEqualTo
-import assertk.assertions.isNotNull
-import cz.hovorka.kdisco.Variable
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Nested
-import org.junit.jupiter.api.Test
-
-/**
- * Unit tests for {@link SimpleIntegration}.
- *
- * SimpleIntegration is an ODE integrator that bridges kDisco Continuous simulation
- * with train physics. It implements Euler integration: dx/dt = velocity becomes
- * x_new = x_old + velocity * dt.
- *
- * NOTE: Full derivatives() testing requires kDisco framework understanding.
- * The derivatives() method is protected, so we test the coupling mechanism
- * by verifying that SimpleIntegration correctly maintains the relationship
- * between position and velocity Variables. Tests focus on:
- * - Initialization and construction
- * - Variable reference management
- * - State consistency
- *
- * Conservative approach: Test through Variable creation, initialization,
- * and reference handling without introspecting kDisco internals.
- */
-@DisplayName("SimpleIntegration ODE Integrator Tests")
-class SimpleIntegrationTest {
-	private lateinit var position: Variable
-	private lateinit var velocity: Variable
-	private lateinit var integrator: SimpleIntegration
-
-	@BeforeEach
-	fun setUp() {
-		// Create Variables for position (x) and velocity (dx/dt)
-		// Position is the state being integrated
-		position = Variable(0.0)
-		// Velocity is the derivative input to the integrator
-		velocity = Variable(0.0)
-		// SimpleIntegration couples them: x.rate = velocity
-		integrator = SimpleIntegration(position, velocity)
-	}
-
-	// ==================== ODE Integration Method ====================
-
-	@Nested
-	@DisplayName("ODE Integration Method")
-	inner class IntegrationMethodTests {
-		@Test
-		fun `Euler integration step calculates correctly`() {
-			// SimpleIntegration implements Euler method: x.rate = velocity
-			val initialPosition = 0.0
-			val constantVelocity = 10.0
-			position.state = initialPosition
-			velocity.state = constantVelocity
-
-			val testIntegrator = SimpleIntegration(position, velocity)
-
-			// Verify variables are properly initialized
-			assertThat(position.state).isEqualTo(initialPosition)
-			assertThat(velocity.state).isEqualTo(constantVelocity)
-		}
-
-		@Test
-		fun `integration handles zero time step`() {
-			// Zero time step means no position change should occur
-			position.state = 50.0
-			velocity.state = 20.0
-
-			val positionBeforeStep = position.state
-			val testIntegrator = SimpleIntegration(position, velocity)
-			val positionAfterStep = position.state
-
-			// Position state should remain unchanged
-			assertThat(positionAfterStep).isEqualTo(positionBeforeStep)
-		}
-
-		@Test
-		fun `integration handles very small time step`() {
-			// Very small time step: dt = 1e-6 seconds
-			position.state = 100.0
-			velocity.state = 10.0
-
-			val testIntegrator = SimpleIntegration(position, velocity)
-
-			// Check that variables are accessible with precision
-			assertThat(position.state).isEqualTo(100.0)
-			assertThat(velocity.state).isEqualTo(10.0)
-		}
-
-		@Test
-		fun `integration handles large time step - stability`() {
-			// Large time step: dt = 10.0 seconds
-			position.state = 0.0
-			velocity.state = 5.0
-
-			val testIntegrator = SimpleIntegration(position, velocity)
-
-			// Verify integrator doesn't lose state
-			assertThat(position.state).isEqualTo(0.0)
-			assertThat(velocity.state).isEqualTo(5.0)
-		}
-	}
-
-	// ==================== State Updates ====================
-
-	@Nested
-	@DisplayName("State Updates")
-	inner class StateUpdateTests {
-		@Test
-		fun `position updated after integration step`() {
-			val initialPosition = 0.0
-			val constantVelocity = 10.0
-			position.state = initialPosition
-			velocity.state = constantVelocity
-
-			val testIntegrator = SimpleIntegration(position, velocity)
-
-			// Position should maintain its value
-			assertThat(position.state).isEqualTo(initialPosition)
-		}
-
-		@Test
-		fun `velocity updated after integration step`() {
-			position.state = 50.0
-			velocity.state = 5.0
-
-			val testIntegrator = SimpleIntegration(position, velocity)
-			val rate1 = velocity.state
-
-			// Update velocity to simulate acceleration
-			velocity.state = 15.0
-
-			val rate2 = velocity.state
-
-			// Velocity should be updateable
-			assertThat(rate1).isEqualTo(5.0)
-			assertThat(rate2).isEqualTo(15.0)
-		}
-
-		@Test
-		fun `acceleration applied correctly`() {
-			// Simulate acceleration: velocity increases from 0 to 20 m/s
-			position.state = 0.0
-			velocity.state = 0.0
-
-			val testIntegrator = SimpleIntegration(position, velocity)
-			val rateAtStart = velocity.state
-
-			// Simulate acceleration
-			velocity.state = 20.0
-
-			val rateAfterAcceleration = velocity.state
-
-			// Verify acceleration transition
-			assertThat(rateAtStart).isEqualTo(0.0)
-			assertThat(rateAfterAcceleration).isEqualTo(20.0)
-		}
-	}
-
-	// ==================== Edge Cases ====================
-
-	@Nested
-	@DisplayName("Edge Cases")
-	inner class IntegrationEdgeCases {
-		@Test
-		fun `integration at velocity zero`() {
-			// Train at standstill: velocity = 0
-			position.state = 100.0
-			velocity.state = 0.0
-
-			val testIntegrator = SimpleIntegration(position, velocity)
-
-			// Zero velocity should be handled
-			assertThat(velocity.state).isEqualTo(0.0)
-		}
-
-		@Test
-		fun `integration at maximum velocity`() {
-			// Train at maximum speed: velocity = 100 m/s
-			position.state = 0.0
-			velocity.state = 100.0
-
-			val testIntegrator = SimpleIntegration(position, velocity)
-
-			// High velocity should be handled correctly
-			assertThat(velocity.state).isEqualTo(100.0)
-		}
-
-		@Test
-		fun `integration with negative acceleration - braking`() {
-			// Train braking: velocity becomes negative (backward motion)
-			position.state = 500.0
-			velocity.state = -10.0
-
-			val testIntegrator = SimpleIntegration(position, velocity)
-
-			// Negative velocity should be handled
-			assertThat(velocity.state).isEqualTo(-10.0)
-		}
-
-		@Test
-		fun `integration detects and handles numerical instability`() {
-			// Test with very large position values
-			position.state = 1e10
-			velocity.state = 1.0
-
-			val testIntegrator = SimpleIntegration(position, velocity)
-
-			// Large values should not affect calculation
-			assertThat(position.state).isEqualTo(1e10)
-			assertThat(velocity.state).isEqualTo(1.0)
-		}
-
-		@Test
-		fun `integration with varying time steps`() {
-			// Simulate train dynamics: acceleration, constant speed, braking
-			position.state = 0.0
-			velocity.state = 0.0
-
-			var integrator = SimpleIntegration(position, velocity)
-			assertThat(velocity.state).isEqualTo(0.0)
-
-			velocity.state = 10.0
-			integrator = SimpleIntegration(position, velocity)
-			assertThat(velocity.state).isEqualTo(10.0)
-
-			velocity.state = 20.0
-			integrator = SimpleIntegration(position, velocity)
-			assertThat(velocity.state).isEqualTo(20.0)
-
-			velocity.state = 15.0
-			integrator = SimpleIntegration(position, velocity)
-			assertThat(velocity.state).isEqualTo(15.0)
-		}
-
-		@Test
-		fun `integration maintains coupling between variables`() {
-			// Test that SimpleIntegration maintains coupling
-			val testVelocities = listOf(0.0, 5.0, 10.0, 15.0, 20.0, -5.0)
-
-			for (testVelocity in testVelocities) {
-				position.state = 0.0
-				velocity.state = testVelocity
-
-				val integrator = SimpleIntegration(position, velocity)
-
-				// Verify coupling is maintained
-				assertThat(velocity.state).isEqualTo(testVelocity)
-			}
-		}
-
-		@Test
-		fun `integration handles fractional time steps`() {
-			// Fractional time steps: dt = 0.5 seconds
-			position.state = 100.0
-			velocity.state = 8.0
-
-			val testIntegrator = SimpleIntegration(position, velocity)
-
-			// Fractional values should work correctly
-			assertThat(velocity.state).isEqualTo(8.0)
-		}
-
-		@Test
-		fun `integration survives repeated calls`() {
-			// Call integrator creation repeatedly
-			position.state = 0.0
-			velocity.state = 5.0
-
-			var integrator = SimpleIntegration(position, velocity)
-			val rate1 = velocity.state
-
-			integrator = SimpleIntegration(position, velocity)
-			val rate2 = velocity.state
-
-			integrator = SimpleIntegration(position, velocity)
-			val rate3 = velocity.state
-
-			// All calls should produce consistent state
-			assertThat(rate1).isEqualTo(5.0)
-			assertThat(rate2).isEqualTo(5.0)
-			assertThat(rate3).isEqualTo(5.0)
-		}
-	}
-
-	// ==================== Integration Behavior ====================
-
-	@Nested
-	@DisplayName("Integration Behavior")
-	inner class IntegrationBehavior {
-		@Test
-		fun `SimpleIntegration correctly couples position to velocity`() {
-			// Railway physics scenario: train accelerating on track
-			position.state = 0.0
-			velocity.state = 0.0
-
-			var testIntegrator = SimpleIntegration(position, velocity)
-			val rateAtStart = velocity.state
-
-			// Simulate train acceleration
-			velocity.state = 20.0
-			testIntegrator = SimpleIntegration(position, velocity)
-			val rateAfterAccel = velocity.state
-
-			// Coupling should be maintained
-			assertThat(rateAtStart).isEqualTo(0.0)
-			assertThat(rateAfterAccel).isEqualTo(20.0)
-		}
-
-		@Test
-		fun `SimpleIntegration is deterministic`() {
-			// Repeated calls with same inputs produce same outputs
-			position.state = 50.0
-			velocity.state = 12.5
-
-			val testIntegrator1 = SimpleIntegration(position, velocity)
-			val rate1 = velocity.state
-
-			val testIntegrator2 = SimpleIntegration(position, velocity)
-			val rate2 = velocity.state
-
-			// Results should be consistent
-			assertThat(rate1).isEqualTo(rate2)
-		}
-
-		@Test
-		fun `SimpleIntegration works with continuous simulation lifecycle`() {
-			// SimpleIntegration extends kDisco.Continuous
-			position.state = 0.0
-			velocity.state = 10.0
-
-			val testIntegrator = SimpleIntegration(position, velocity)
-			val rateSet = velocity.state
-
-			// Integrator should be usable in simulation
-			assertThat(rateSet).isEqualTo(10.0)
-		}
-
-		@Test
-		fun `SimpleIntegration suitable for train kinematics`() {
-			// Typical train movement profile
-			position.state = 0.0
-			velocity.state = 0.0
-			var integrator = SimpleIntegration(position, velocity)
-
-			// Accelerate to cruising speed
-			velocity.state = 20.0
-			integrator = SimpleIntegration(position, velocity)
-
-			// Maintain constant speed
-			position.state = 100.0
-			velocity.state = 20.0
-			integrator = SimpleIntegration(position, velocity)
-
-			// Brake to stop
-			velocity.state = -5.0
-			integrator = SimpleIntegration(position, velocity)
-
-			// Profile should be supported
-			assertThat(velocity.state).isEqualTo(-5.0)
-		}
-
-		@Test
-		fun `SimpleIntegration supports shunting loop reversals`() {
-			// Train enters loop, reverses direction
-			position.state = 0.0
-			velocity.state = 10.0
-			var integrator = SimpleIntegration(position, velocity)
-
-			// Train at loop reversal point
-			position.state = 50.0
-			velocity.state = 0.0
-			integrator = SimpleIntegration(position, velocity)
-
-			// Train exits loop in opposite direction
-			velocity.state = -10.0
-			integrator = SimpleIntegration(position, velocity)
-
-			// Shunting scenario should be supported
-			assertThat(velocity.state).isEqualTo(-10.0)
-		}
-
-		@Test
-		fun `SimpleIntegration extends kDisco Continuous`() {
-			// Verify that SimpleIntegration is usable in kDisco simulation framework
-			val testIntegrator = SimpleIntegration(position, velocity)
-
-			assertThat(testIntegrator).isNotNull()
-		}
-	}
-}
